+++ vanzare/core/settings/base.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'widget_tweaks',
     'django.contrib.admin',
     'django.contrib.auth',
+    'producto',
 ]
 
 MIDDLEWARE = [
