+++ vanzare/core/urls.py
@@ -19,5 +19,6 @@ urlpatterns = [
     url(r'^accounts/reset/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$', auth_views.password_reset_confirm, name='password_reset_confirm'),
     url(r'^accounts/reset/done/$', auth_views.password_reset_complete, name='password_reset_complete'),
     url(r'^', include('facturacion.urls')),
+    url(r'^', include('producto.urls')),
     url(r'^admin/', admin.site.urls),
 ]
