+++ vanzare/facturacion/urls.py
@@ -24,8 +24,5 @@ urlpatterns = [
     url(r'^clientes/listar/$', views.clientesListar, name='clientes-listar'),
     url(r'^clientes/crear/$', views.clientesCrear, name='clientes-crear'),
     url(r'^clientes/editar/(?P<pk>[0-9]+)/$', views.clientesEditar, name='clientes-editar'),
-    url(r'^productos/listar/$', views.productosListar, name='productos-listar'),
-    url(r'^productos/crear/$', views.productosCrear, name='productos-crear'),
-    url(r'^productos/editar/(?P<pk>[0-9]+)/$', views.productosEditar, name='productos-editar'),
     url(r'^troregis/$', views.registroUsuario, name='registro'),
 ]
